Split up test_tomography into multiple files
